Update Actions folder
